Refactor property types in SkuPrice and ProductRepository.php

+++ src/Entities/Catalog/SkuPrice.php
@@ -7,9 +7,9 @@ use Orkestra\Entities\AbstractEntity;
 /**
  * @property-read string $itemId
  * @property-read float|null $listPrice
- * @property-read int $costPrice
+ * @property-read float $costPrice
  * @property-read float $markup
- * @property-read int $basePrice
+ * @property-read float $basePrice
  * @property-read array<int,array{tradePolicyId:string,value:float,listPrice:float|null,minQuantity:int,dateRange:array{from:string,to:string}}> $fixedPrices
  */
 class SkuPrice extends AbstractEntity
@@ -20,9 +20,9 @@ class SkuPrice extends AbstractEntity
 	public function __construct(
 		protected string $itemId,
 		protected ?float $listPrice,
-		protected int $costPrice,
+		protected float $costPrice,
 		protected float $markup,
-		protected int $basePrice,
+		protected float $basePrice,
 		protected array $fixedPrices,	
 	) {
 		//
